node: Remove the legacy Handler interface
All production node consumers now use StateDB, and the fake datapath reads the
node table directly. Remove Handler subscription and dispatch from NodeManager
along with its callback-based background validation loop.

NodeManager temporarily retains its cache and producer-facing APIs. Follow-up
patches can turn those APIs into thin Writer and table facades before migrating
producers independently.

REVIEW: NodeManager event health scopes are removed because they only reported
synchronous Handler callback failures. Reconciler and observer health now owns
consumer failures.

AIL:3
Related: https://github.com/cilium/cilium/issues/41744
